requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
filelock==3.25.2
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.2
Werkzeug==3.1.7
pytest==9.0.2
pytest-mock==3.15.1
